Capture/Ball Fetch Fixed
gLastUsedBall increased to u16 to fix overflow.
gBattleResults.CatchAttempts set to use ItemIdToBallId like it's supposed to.
Strange Ball exceptions removed since the above two were fixed.

+++ include/battle.h
@@ -1123,7 +1123,7 @@ extern u8 gHealthboxSpriteIds[MAX_BATTLERS_COUNT];
 extern u8 gMultiUsePlayerCursor;
 extern u8 gNumberOfMovesToChoose;
 extern bool8 gHasFetchedBall;
-extern u8 gLastUsedBall;
+extern u16 gLastUsedBall;
 extern u16 gLastThrownBall;
 extern u16 gBallToDisplay;
 extern bool8 gLastUsedBallMenuPresent;
